backend/github_issues_ingestion/requirements.txt
# Requirements for GitHub Issues Ingestion System
# These should already be installed in the parent project

# GitHub API
requests>=2.31.0

# Azure OpenAI
openai>=1.0.0

# Vector Database
pymilvus>=2.3.0

# Environment variables
python-dotenv>=1.0.0

# Optional: For better date handling
python-dateutil>=2.8.2

